Zdoom extended nodes. Code generator is implemented and added to Makefile (via go generate, also see level.go). znodegen.go is a generated file, not for manual editing. Version renamed to v0.75 (from v0.75a)

+++ gamespec.go
@@ -387,7 +387,7 @@ type ZdoomNode_Seg struct {
 	StartVertex uint32
 	EndVertex   uint32
 	Linedef     uint16
-	Flip        int16
+	Flip        byte
 }
 
 // ZdoomNode_node is not defined - it is same as DeepNode
